refactor(mcp): consolidate ParameterDisplay into RequestDetailPanel #371
- Move ParameterDisplay implementation into RequestDetailPanel file
- Simplify response duration label in ResponseDetailPanel
- Remove standalone ParameterDisplay.kt file
